Change from uint64_t to size_t default index type

+++ include/grb/containers/matrix.hpp
@@ -11,7 +11,7 @@
 namespace grb {
 
 template <typename T,
-          typename I = std::uint64_t,
+          typename I = std::size_t,
           typename Allocator = std::allocator<T>>
 class matrix {
 public:
